Fix message shown when brew is not present in Darwin
The message `not='' fount=''` was shown when brew is not installed. This is because the
message "brew not found" is output but only the first word "brew" is used. To
fix this, awk is used as a workaround to only get the first word. I am sure
there is a better shell script way of doing this, but this is a workaround.