Fix error message about undefined variable `name'
This error message is fixed by lazy-loading neocomplete.vim, and therefore,
letting vimproc.vim load first. This issue only appeared when resolving
neocomplete.vim dependencies (which is only vimproc.vim). The cause of this is
thought to be the fact that both neocomplete.vim and vimproc.vim are
`NeoBundle'ed instead of one being `NeoBundleLazy'ed.