Fix bug in malloc snippet where wrong data type is used
In order to get around this issue, sizeof(${1}) is filled in first, then the
multiplier, and finally what to cast to (with the default value set to pointer
to $1). Ideally, the snippet will be something like:

(${1:#:type)malloc(sizeof(${2:`substitute($1, '\*$', '', '')`}) * ${3});${0}

but as of 5/11/2015, this feature is labeled as "hard to implement" in
neosnippet.vim issue #270.