Fix bug where vim-clang is sourced when vim-marching is available
This bug is due to the way in which TOML files are processed. neobundle.vim
reads in the entire TOML file using the parser, then later on reflects that to
its list of bundles. The order in which the plugins are reflected is
undefined, and therefore, neobundle#is_installed('vim-marching') returns
undefined value of 1 or 0. vimproc.vim is used as an implicit indication of
whether vim-marching can be used or not. This works because vimproc.vim is
read in from a different file, and therefore, is guaranteed that
neobundle#is_installed() returns the correct value. Care must be taken when
changing the order in which vimproc.vim is read.