Add variable to control whether to use vim-monster
This only has effect when set to 0 since vim-monster is used by default for
Ruby files. g:use_monster can be set to 0 when there is need to explicitly
disable vim-monster (such as when trying out new completion plugins). Note
that g:use_monster = 0 does NOT imply that rsense will be used. Set
g:use_rsense = 1 when needing to do so.