Emacs: use C-x w for ace-window and C-x o for other-window
The rationale for this change is that C-x {o,O} smartrep.el group defines
behaviors for switching windows. Thus, it is more natural to add the
ace-window key binding to this group rather than randomly binding M-n. It also
makes sense to simply add a different suffix (in this case 'w') rather than
overriding C-x o for no reason.