Use alpine in workspace 6
This change assigns workspace 6 to alpine, with notifications in xmobar. The
notifications utilize the urgency hook for xmonad and catches the bell emitted
by alpine on receiving new messages. Workspace 6 is assigned because it is the
farthest key from the home position, thus, prevents accidentally switching to
the workspace and dismissing the notification.