Add option to preserve workspace paths when moving out
This option is enabled by default in the PRESERVE_LAST_WS variable. If
enabled, the workspace paths are preserved until being cd-ed into a different
workspace, in which case the paths are rolled back and the setup.zsh in the
new workspace is sourced. If disabled, the paths are reset when leaving the
workspace root directory to whatever state it was before coming into that
workspace.